third_party/blink/web_tests/platform/linux-chrome/external/wpt/speculation-rules/prerender/response-code-successful_code=232-expected.txt

This is a testharness.js-based test.
[FAIL] Responses with code 232 should be activated
  assert_equals: expected "activated" but got "discarded"
Harness: the test ran to completion.
